Simplify theming engine with templates and external tools

+++ user/resources/theme/theme.sh
@@ -1,21 +1,23 @@
 #!/usr/bin/env bash
-rm ~/.cache/wallpaper
-# wal -i $1
-# wpg -s $1 --alpha 84
-ln -s $1 ~/.cache/wallpaper
-./move-colors-kde.sh Pywal
-./apply-firefox-css.sh
-sed "s/Opacity=1/Opacity=0.84/g" $HOME/.cache/wal/colors-konsole.colorscheme >$HOME/.local/share/konsole/colors-konsole.colorscheme
+(
+if [ -e "$1.scheme" ]; then
+  echo test
+  # pass
+else
+  rm ~/.cache/wallpaper
+  ln -s $1 ~/.cache/wallpaper
+  feh --bg-fill $1
+  kde-material-you-colors -wal -kp Pywal -ko 84 --on-change-hook "kde-material-you-colors --stop"
+  if [ ! -d "~/.local/share/themes/FlatColor" ]; then
+    wpg-install.sh -g
+  fi
+  wpg -i ~/.cache/wallpaper ~/.cache/wal/colors.json
+  wpg -s ~/.cache/wallpaper
+fi
 cp $HOME/.cache/wal/colors-albert.qss $HOME/.local/share/albert/widgetsboxmodel/themes/Pywal.qss
-cp ~/.cache/wal/gtk2 ~/.gtkrc-2.0
-cp ~/.cache/wal/gtk2 ~/.config/gtkrc-2.0
-cp ~/.cache/wal/gtk2 ~/.config/gtkrc
-cp ~/.cache/wal/gtk2 ~/.local/share/themes/FlatColor/gtk-2.0/gtkrc
-cp ~/.cache/wal/gtk3.0 ~/.config/gtk-3.0/gtk.css
-cp ~/.cache/wal/gtk3.0 ~/.local/share/themes/FlatColor/gtk-3.0/gtk.css
-cp ~/.cache/wal/gtk3.20 ~/.local/share/themes/FlatColor/gtk-3.20/gtk.css
-cp ~/.cache/wal/gtk3.20 ~/.config/gtk-4.0/gtk.css
-feh --bg-fill $1
+albert restart
+sleep 1
+sed -i 's/BackgroundNormal=#/BackgroundNormal=#D6/g' ~/.local/share/color-schemes/MaterialYouDark.colors
 lookandfeeltool -a seaglass
 systemctl --user restart picom
-
+) &
